Changed Chapter PDF compiling

+++ nodePrint/nodePrint.js
@@ -1168,13 +1168,13 @@ puppeteer.launch({
 				
 				return {filename: escapedURL + '.pdf'};
 			}
-/*			else if (!isNoCache && working[escapedURL]) { //another thread is already working
-				eventEmitter.on(escapedURL, () => {
-					console.log(`DUPE   ${ip} ${url}`);
-					// staticFileServer.serveFile('../PDF/' + escapedURL + '.pdf', 200, {'cache-control': 'no-cache'}, request, response);
-				});
-				return false;
-			}*/
+			/*			else if (!isNoCache && working[escapedURL]) { //another thread is already working
+							eventEmitter.on(escapedURL, () => {
+								console.log(`DUPE   ${ip} ${url}`);
+								// staticFileServer.serveFile('../PDF/' + escapedURL + '.pdf', 200, {'cache-control': 'no-cache'}, request, response);
+							});
+							return false;
+						}*/
 			
 			const start = performance.now();
 			console.log(`NEW    ${ip} ${url}`);
@@ -1508,6 +1508,7 @@ puppeteer.launch({
 			console.log(`Getting LibreText ${options.index ? `[${options.index}] ` : ''}${current.title}`);
 			const zipFilename = `${current.subdomain}-${current.id}`;
 			const thinName = md5(zipFilename).slice(0, 6);
+			const hasCoverpage = current.tags.includes('coverpage:yes');
 			let privatePages = [];
 			
 			//Try to get special files
@@ -1517,7 +1518,7 @@ puppeteer.launch({
 			let TOCIndex = ++totalIndex;
 			
 			async function getMatter(text) {
-				if (!current.tags.includes('coverpage:yes')) {
+				if (!hasCoverpage) {
 					console.log(`Skipping ${text} matter`);
 					return [];
 				}
@@ -1671,7 +1672,8 @@ puppeteer.launch({
 			
 			await fs.emptyDir(`./PDF/Letter/libretexts/${zipFilename}`);
 			await fs.emptyDir(`./PDF/Letter/Finished/${zipFilename}`);
-			await fs.ensureDir(`./PDF/Letter/Finished/${zipFilename}/Publication`);
+			if (hasCoverpage)
+				await fs.ensureDir(`./PDF/Letter/Finished/${zipFilename}/Publication`);
 			await fs.emptyDir(`./PDF/Letter/order/${thinName}/`);
 			
 			// await fs.emptyDir(`./PDF/A4/libretexts/${zipFilename}`);
@@ -1797,10 +1799,11 @@ puppeteer.launch({
 				// await fs.copy(`./PDF/A4/${filename}`, `./PDF/A4/libretexts/${zipFilename}/${filenamify('00000:A Cover.pdf')}`);
 				// await fs.copy(`./PDF/A4/${filename}`, `./PDF/A4/order/${thinName}/${`0`.padStart(3, '0')}.pdf`);
 				
-				
-				let dest = await getThinCC(current, `./PDF/Letter/Finished/${zipFilename}/LibreText.imscc`);
-				// if (dest)
-				// await fs.copy(`./PDF/Letter/Finished/${zipFilename}/LibreText.imscc`, `./PDF/A4/Finished/${zipFilename}/LibreText.imscc`);
+				if (hasCoverpage) {
+					let dest = await getThinCC(current, `./PDF/Letter/Finished/${zipFilename}/LibreText.imscc`);
+					// if (dest)
+					// await fs.copy(`./PDF/Letter/Finished/${zipFilename}/LibreText.imscc`, `./PDF/A4/Finished/${zipFilename}/LibreText.imscc`);
+				}
 				let files = (await fs.readdir(`./PDF/Letter/order/${thinName}`)).map((file) => `./PDF/Letter/order/${thinName}/${file}`);
 				// let filesA4 = (await fs.readdir(`./PDF/A4/order/${thinName}`)).map((file) => `./PDF/A4/order/${thinName}/${file}`);
 				console.log(`Merging${options.index ? ` [${options.index}]` : ''}`);
@@ -1819,17 +1822,21 @@ puppeteer.launch({
 					if (files && files.length > 2) {
 						await merge(files, `./PDF/Letter/Finished/${zipFilename}/Full.pdf`, {maxBuffer: 1024 * 10000000});
 						// await merge(filesA4, `./PDF/A4/Finished/${zipFilename}/Full.pdf`, {maxBuffer: 1024 * 10000000});
-						await merge(files.slice(0, 10), `./PDF/Letter/Finished/${zipFilename}/Preview.pdf`, {maxBuffer: 1024 * 10000000});
-						files.shift();
-						// filesA4.shift();
-						await merge(files, `./PDF/Letter/Finished/${zipFilename}/Publication/Content.pdf`, {maxBuffer: 1024 * 10000000});
-						// await merge(filesA4, `./PDF/A4/Finished/${zipFilename}/Publication/Content.pdf`, {maxBuffer: 1024 * 10000000});
+						if (hasCoverpage) {
+							await merge(files.slice(0, 10), `./PDF/Letter/Finished/${zipFilename}/Preview.pdf`, {maxBuffer: 1024 * 10000000});
+							files.shift();
+							// filesA4.shift();
+							await merge(files, `./PDF/Letter/Finished/${zipFilename}/Publication/Content.pdf`, {maxBuffer: 1024 * 10000000});
+							// await merge(filesA4, `./PDF/A4/Finished/${zipFilename}/Publication/Content.pdf`, {maxBuffer: 1024 * 10000000});
+						}
 					}
 					else {
 						await fs.copy(files[0], `./PDF/Letter/Finished/${zipFilename}/Full.pdf`);
-						await fs.copy(files[0], `./PDF/Letter/Finished/${zipFilename}/Publication/Content.pdf`);
-						// await fs.copy(filesA4[0], `./PDF/Letter/Finished/${zipFilename}/Full.pdf`);
-						// await fs.copy(filesA4[0], `./PDF/A4/Finished/${zipFilename}/Publication/Content.pdf`);
+						if (hasCoverpage) {
+							await fs.copy(files[0], `./PDF/Letter/Finished/${zipFilename}/Publication/Content.pdf`);
+							// await fs.copy(filesA4[0], `./PDF/Letter/Finished/${zipFilename}/Full.pdf`);
+							// await fs.copy(filesA4[0], `./PDF/A4/Finished/${zipFilename}/Publication/Content.pdf`);
+						}
 					}
 					console.log(`Done Merging${options.index ? ` [${options.index}]` : ''}`);
 				} catch (e) {
@@ -1878,50 +1885,53 @@ puppeteer.launch({
 					// await fs.copy(`./PDF/A4/${filename}`, `./PDF/A4/Finished/${zipFilename}/Publication/Cover_CoilBound.pdf`);
 				}
 				//save log of private pages
-				if (privatePages.length) {
+				if (hasCoverpage && privatePages.length) {
 					await fs.writeFile(`./PDF/Letter/Finished/${zipFilename}/Publication/Notice_Private_Pages.txt`, privatePages.join('\n'));
 					// await fs.writeFile(`./PDF/A4/Finished/${zipFilename}/Publication/Notice_Private_Pages.txt`, privatePages);
 				}
 				
-				console.log(`Zipping${options.index ? ` [${options.index}]` : ''}`);
-				let individualZIP = new JSZip();
-				// let individualZIPA4 = new JSZip();
-				let PublicationZIP = new JSZip();
-				// let PublicationZIPA4 = new JSZip();
-				files = await fs.readdir('./PDF/Letter/libretexts/' + zipFilename);
-				// filesA4 = await fs.readdir('./PDF/A4/libretexts/' + zipFilename);
-				for (let i = 0; i < files.length; i++) {
-					individualZIP.file(files[i], await fs.readFile(`./PDF/Letter/libretexts/${zipFilename}/${files[i]}`));
-					// individualZIPA4.file(filesA4[i], await fs.readFile(`./PDF/A4/libretexts/${zipFilename}/${filesA4[i]}`));
-				}
-				files = await fs.readdir(`./PDF/Letter/Finished/${zipFilename}/Publication`);
-				// filesA4 = await fs.readdir(`./PDF/A4/Finished/${zipFilename}/Publication`);
-				for (let i = 0; i < files.length; i++) {
-					PublicationZIP.file(files[i], await fs.readFile(`./PDF/Letter/Finished/${zipFilename}/Publication/${files[i]}`));
-					// PublicationZIPA4.file(filesA4[i], await fs.readFile(`./PDF/A4/Finished/${zipFilename}/Publication/${filesA4[i]}`));
-				}
-				
-				
-				await saveAs(individualZIP, `./PDF/Letter/Finished/${zipFilename}/Individual.zip`);
-				// await saveAs(individualZIPA4, `./PDF/A4/Finished/${zipFilename}/Individual.zip`);
-				await saveAs(PublicationZIP, `./PDF/Letter/Finished/${zipFilename}/Publication.zip`);
-				
-				// await saveAs(PublicationZIPA4, `./PDF/A4/Finished/${zipFilename}/Publication.zip`);
-				
-				async function saveAs(zip, destination) {
-					let result = await zip.generateAsync({type: "nodebuffer"});
-					await fs.writeFile(destination, result);
-				}
-				
-				if (failed) {
-					console.error(`Merge Failed, clearing cache${options.index ? ` [${options.index}]` : ''}`);
-					await async.mapLimit(originalFiles, 10, async (filename) => {
-						await fs.remove(`./PDF/Letter/${filename}`);
-						await fs.remove(`./PDF/Letter/Margin/${filename}`);
-						// await fs.remove(`./PDF/A4/${filename}`);
-						// await fs.remove(`./PDF/A4/Margin/${filename}`);
-					});
-					console.error(`Cache ${zipFilename} cleared${options.index ? ` [${options.index}]` : ''}`);
+				//creating zip files
+				if (hasCoverpage) {
+					console.log(`Zipping${options.index ? ` [${options.index}]` : ''}`);
+					let individualZIP = new JSZip();
+					// let individualZIPA4 = new JSZip();
+					let PublicationZIP = new JSZip();
+					// let PublicationZIPA4 = new JSZip();
+					files = await fs.readdir('./PDF/Letter/libretexts/' + zipFilename);
+					// filesA4 = await fs.readdir('./PDF/A4/libretexts/' + zipFilename);
+					for (let i = 0; i < files.length; i++) {
+						individualZIP.file(files[i], await fs.readFile(`./PDF/Letter/libretexts/${zipFilename}/${files[i]}`));
+						// individualZIPA4.file(filesA4[i], await fs.readFile(`./PDF/A4/libretexts/${zipFilename}/${filesA4[i]}`));
+					}
+					files = await fs.readdir(`./PDF/Letter/Finished/${zipFilename}/Publication`);
+					// filesA4 = await fs.readdir(`./PDF/A4/Finished/${zipFilename}/Publication`);
+					for (let i = 0; i < files.length; i++) {
+						PublicationZIP.file(files[i], await fs.readFile(`./PDF/Letter/Finished/${zipFilename}/Publication/${files[i]}`));
+						// PublicationZIPA4.file(filesA4[i], await fs.readFile(`./PDF/A4/Finished/${zipFilename}/Publication/${filesA4[i]}`));
+					}
+					
+					
+					await saveAs(individualZIP, `./PDF/Letter/Finished/${zipFilename}/Individual.zip`);
+					// await saveAs(individualZIPA4, `./PDF/A4/Finished/${zipFilename}/Individual.zip`);
+					await saveAs(PublicationZIP, `./PDF/Letter/Finished/${zipFilename}/Publication.zip`);
+					
+					// await saveAs(PublicationZIPA4, `./PDF/A4/Finished/${zipFilename}/Publication.zip`);
+					
+					async function saveAs(zip, destination) {
+						let result = await zip.generateAsync({type: "nodebuffer"});
+						await fs.writeFile(destination, result);
+					}
+					
+					if (failed) {
+						console.error(`Merge Failed, clearing cache${options.index ? ` [${options.index}]` : ''}`);
+						await async.mapLimit(originalFiles, 10, async (filename) => {
+							await fs.remove(`./PDF/Letter/${filename}`);
+							await fs.remove(`./PDF/Letter/Margin/${filename}`);
+							// await fs.remove(`./PDF/A4/${filename}`);
+							// await fs.remove(`./PDF/A4/Margin/${filename}`);
+						});
+						console.error(`Cache ${zipFilename} cleared${options.index ? ` [${options.index}]` : ''}`);
+					}
 				}
 			}
 			const end = performance.now();
